Adds FSRS bulk-loading cron job
FSRS data needs to be downloaded in chunks of 500 entries. To get us started,
this installs cron to the job manager and sets up a cron job to execute the
`loadFSRS` script. It's currently appending the results (a line or two per
run) to a tmp file, but that's only a temporary solution.